test(core): extract inline test block in changelog.rs to sibling tests.rs
Workspace policy (Makefile:412 lint-tests-layout) forbids inline
#[cfg(test)] mod tests {...} blocks. Move the 6 tests into
schema/changelog/tests.rs.

cargo nextest -p fraiseql-core --lib changelog → 11/11 pass.
